Formated result. Fixed bug caused by not completing switch from int to double. Using Double.parseDouble#

+++ src/edu/uib/info231/model/ConverterModel.java
@@ -9,7 +9,7 @@ public class ConverterModel {
 		this.to = to;
 	}
 	
-	public double convert(int fromAmount) {
+	public double convert(double fromAmount) {
 		return (from.getValue() / to.getValue()) * fromAmount;
 	}
 }
